+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-08-17
+
+### Added
+- Added extended output in debug mode [#5](https://github.com/sajya/server/pull/5)
+
+### Fixed
+- Fixed correct header output [#5](https://github.com/sajya/server/pull/5)
+
 ## 0.0.2 - 2020-03-23
 
 ### Fixed